requirements.txt
crewai==0.11.0
langchain-community==0.0.14
python-dotenv
google-search-results==2.1.0
google-api-python-client==2.114.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
beautifulsoup4==4.12.3
tavily-python==0.3.1
langchain_google_genai
sympy
fpdf==1.7.2 
colorama
Markdown
serpapi
requests
Pillow
Flask